+++ .gitignore
@@ -1,0 +1,3 @@
+game_of_life
+main.o
+gol.o
